Lab 3 - zrobione zadania 1-4

+++ solution/lab2/zadanie1.py
@@ -29,15 +29,16 @@ source = "../../resources/lab1/files"
 for f in os.listdir(source):
     # Otwieramy plik. os.sep jest to zmienna globalna przechowujaca znak okreslajacy rozdzielenie folderow w sciezkach
     # w zależnosci od systemu na jakim jest uruchomiony kod(linux: /, windows: \)
-    with open(source+os.sep+f, "r") as rf:
-        # Wyciagamy nazwe katalogu w ktorym powinny byc umieszczone pliki.
-        #   metoda replace w klasie string zamienia znaki ze spacjji na pusty znak i zwraca nowy string
-        #   metoda split w klasie string zwraca tablice stringow wynikajaca z podzialu stringa za pomoca :
-        dirname = rf.read().replace(' ', '').split(':')[1]
+    if os.path.isfile(source+os.sep+f):
+        with open(source+os.sep+f, "r") as rf:
+            # Wyciagamy nazwe katalogu w ktorym powinny byc umieszczone pliki.
+            #   metoda replace w klasie string zamienia znaki ze spacjji na pusty znak i zwraca nowy string
+            #   metoda split w klasie string zwraca tablice stringow wynikajaca z podzialu stringa za pomoca :
+            dirname = rf.read().replace(' ', '').split(':')[1]
 
-    # sprawadzamy czy folder o podanej nazwie istnieje. Jeśli nie to go tworzymy.
-    if not os.path.exists(dirname):
-        os.makedirs(dirname)
+        # sprawadzamy czy folder o podanej nazwie istnieje. Jeśli nie to go tworzymy.
+        if not os.path.exists(dirname):
+            os.makedirs(dirname)
 
-    # Kopiujemy plik z poczatkowego katalogu do wskazanego przez teks znajdujacy sie pliku
-    shutil.copyfile(source+os.sep+f, dirname+os.sep+f)
+        # Kopiujemy plik z poczatkowego katalogu do wskazanego przez teks znajdujacy sie pliku
+        shutil.copyfile(source+os.sep+f, dirname+os.sep+f)
